Extract aspect, result, and complement composition
