+++ src/plugin-entry.js
@@ -969,6 +969,11 @@ const {
   npLicenseMetadata,
 });
 
+const resultComplementModule = require("./parser/detectors/complements/result");
+const resultFramePartClone = resultComplementModule.createResultFramePartClone({
+  firstToken, flattenSurface, parserInactiveTokenClone, token,
+});
+
 const createNpDetectors = require("./parser/detectors/np/core");
 const {
   approximateQuantityFallback,
@@ -2005,96 +2010,14 @@ function transparentTopicContentFromNodes(nodes) {
   return categorySubspanFor(compact, ["OvertHeadDemonstrativeClassifierNP", "QuantifiedClassifierNP", "QuantifiedPersonNP", "OrdinalClassifierNP", "DiMarkedNP", "ModifiedNP", "NominalHeadSpan", "CoordinatedNP"]);
 }
 
-function potentialResultVPFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  if (!bareCore.length) return null;
-  let cursor = 0;
-  const subject = nodeCanFillSlot(bareCore[cursor], "subject") ? bareCore[cursor++] : null;
-  const action = bareCore[cursor++];
-  if (!action || !isVerbLike(action) || !isToken(bareCore[cursor], "得")) return null;
-  const markerSource = bareCore[cursor++];
-  const resultNodes = bareCore.slice(cursor);
-  if (!resultNodes.length) return null;
-  if (!resultNodes.some((node) => nodeCanFillSlot(node, "result_complement") || nodeCanFillSlot(node, "completion_marker") || ["完", "到", "掂", "切"].includes(flattenSurface(node)))) return null;
-  const resultHead = resultNodes[0];
-  const objectSource = resultNodes.slice(1);
-  let objectNode = null;
-  if (objectSource.length === 1 && (nodeCanFillSlot(objectSource[0], "object") || nodeCanFillSlot(objectSource[0], "np") || nodeCanFillSlot(objectSource[0], "head_noun"))) {
-    objectNode = objectSource[0].kind === "construction" ? objectSource[0] : (categorySubspanFor(objectSource, ["NominalHeadSpan"]) || objectSource[0]);
-  } else if (objectSource.length > 1) {
-    objectNode = classifierObjectNPFromNodes(objectSource) || categorySubspanFor(objectSource, ["OvertHeadDemonstrativeClassifierNP", "QuantifiedClassifierNP", "QuantifiedPersonNP", "OrdinalClassifierNP", "DiMarkedNP", "ModifiedNP", "NominalHeadSpan"]);
-  }
-  if (objectSource.length && !objectNode) return null;
+const createPotentialResultDetectors = require("./parser/detectors/aspect/potential-result");
+const { potentialResultVPFallback, incompletePotentialResultCandidate } = createPotentialResultDetectors({
+  categorySubspanFor, classifierObjectNPFromNodes, cleanSlots, construction, flattenSurface,
+  isToken, isVerbLike, nodeCanFillSlot, parserInactiveTokenClone, templateDerivedSlots,
+  traceInfo, withoutTrailingParticles,
+});
 
-  const marker = parserInactiveTokenClone(markerSource, {
-    label: "func",
-    pos: "function",
-    syntax: "potential_marker",
-    slots: ["potential_marker"],
-    reason: "Between an action predicate and an overt result complement, 得 is the productive positive potential linker, not a standalone acceptability response.",
-  });
-  const children = [action, marker, resultHead, ...(objectNode ? [objectNode] : [])];
-  const potential = construction("PotentialResultVP", "Potential", children, {
-    slots: cleanSlots(["potential_result_vp", "potential_marker", "result_complement", "vp", "action_vp", "predicate"]),
-    note: "Productive positive potential construction: action + 得 + overt result complement.",
-    trace: traceInfo("generative_template", {
-      construction_type: "PotentialResultVP",
-      template_family: "generative_template",
-      template: ["action_verb!", "potential_marker!", "result_complement!"],
-      assigned_slots: ["action_verb", "potential_marker", "result_complement", ...(objectNode ? ["object"] : [])],
-      surfaces: children.map((node) => flattenSurface(node)),
-      potential_polarity: "positive",
-      not_claims: ["not_acceptability_response", "not_hidden_result_complement"],
-    }),
-  });
-  if (!subject) return potential;
-  const clauseChildren = [subject, potential, ...particles];
-  return construction("SubjectPredicateClause", "SubjPred", clauseChildren, {
-    slots: templateDerivedSlots("SubjectPredicateClause", clauseChildren),
-    note: "Subject plus a productive positive potential predicate.",
-    trace: traceInfo("generative_template", {
-      construction_type: "SubjectPredicateClause",
-      template_family: "generative_template",
-      template: ["subject!", "predicate!", "particle?"],
-      assigned_slots: ["subject", "predicate", ...particles.map(() => "particle")],
-      surfaces: clauseChildren.map((node) => flattenSurface(node)),
-    }),
-  });
-}
 
-function incompletePotentialResultCandidate(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  let cursor = 0;
-  const subject = nodeCanFillSlot(bareCore[cursor], "subject") ? bareCore[cursor++] : null;
-  const action = bareCore[cursor++];
-  const marker = bareCore[cursor++];
-  if (cursor !== bareCore.length || !action || !isVerbLike(action) || !isToken(marker, "得")) return null;
-  const children = [...(subject ? [subject] : []), action, marker, ...particles];
-  return construction("NeedsContext", "needs context", children, {
-    slots: cleanSlots(["needs_context", "review_candidate", "predicate", "problem_span", subject ? "subject" : ""]),
-    note: "Potential/acceptability boundary with no overt result complement.",
-    trace: traceInfo("special_ambiguity_rule", {
-      construction_type: "NeedsContext",
-      predicate_omission_profile: "acceptability_possibility",
-      omission_status: "potential_result_or_contextual_ellipsis_ambiguous",
-      template: ["subject?", "action_verb!", "potential_marker_or_acceptability_predicate!", "particle?"],
-      assigned_slots: [...(subject ? ["subject"] : []), "action_verb", "potential_marker_or_acceptability_predicate", ...particles.map(() => "particle")],
-      surfaces: children.map((node) => flattenSurface(node)),
-      missing_argument_slots: ["result_or_activity_domain"],
-      missing_slot_details: [{ slot: "result_or_activity_domain", license_status: "unresolved" }],
-      complement_type: "result_complement_or_contextual_activity",
-      context_requirement_status: "context_required",
-      antecedent_status: "not_observed",
-      selected_alternative: "underdetermined",
-      subject_status: subject ? "explicit" : "omitted_unlicensed",
-      polarity: "positive",
-      conventionality_status: "context_sensitive",
-      speech_event_use: "not_applicable",
-      semantic_review_flags: ["needs_context_parse", "acceptability_potential_boundary"],
-      not_claims: ["not_clean_acceptability_response", "not_fabricated_result_complement", "not_sentence_specific_surface_rule"],
-    }),
-  });
-}
 
 
 
@@ -2270,78 +2193,23 @@ function interiorExistentialFrameFallback(core) {
 }
 
 
-function resultFramePartClone(node, overrides = {}) {
-  const surface = overrides.surface || flattenSurface(node);
-  const base = firstToken(node) || token(surface);
-  return parserInactiveTokenClone(base, {
-    label: overrides.label || base.label || "func",
-    pos: overrides.pos || (overrides.label === "doing" ? "verb" : overrides.label === "particle" ? "particle" : overrides.label === "when" ? "adverbial" : "function"),
-    syntax: overrides.syntax || base.syntax || "result_frame_part",
-    slots: overrides.slots || [],
-    reason: overrides.reason || "Token is parser-inactive inside a bounded v0.5.34 change/result frame; the parent exposes the result relation while child tokens stay visible.",
-  });
-}
-
-function resultTopicFromNodes(nodes) {
-  const compact = withoutIgnorableSpaceText(nodes || []);
-  if (!compact.length) return null;
-  if (compact.length === 1 && (nodeCanFillSlot(compact[0], "topic") || nodeCanFillSlot(compact[0], "subject") || nodeCanFillSlot(compact[0], "location") || nodeCanFillSlot(compact[0], "np") || nodeCanFillSlot(compact[0], "head_noun"))) return compact[0];
-  const templated = categorySubspanFor(compact, ["OvertHeadDemonstrativeClassifierNP", "QuantifiedClassifierNP", "DiMarkedNP", "OrdinalClassifierNP", "NominalHeadSpan", "CoordinatedNP"]);
-  if (templated && (nodeCanFillSlot(templated, "topic") || nodeCanFillSlot(templated, "subject") || nodeCanFillSlot(templated, "np") || nodeCanFillSlot(templated, "head_noun"))) return templated;
-  return null;
-}
-
-function resultComplementFromNodes(nodes) {
-  const compact = withoutIgnorableSpaceText(nodes || []);
-  if (!compact.length) return null;
-  const nominal = nominalComplementFromNodes(compact);
-  if (nominal) return nominal;
-  if (compact.length === 1 && (nodeCanFillSlot(compact[0], "object") || nodeCanFillSlot(compact[0], "np") || nodeCanFillSlot(compact[0], "head_noun") || nodeCanFillSlot(compact[0], "location"))) return compact[0];
-  return null;
-}
+const {
+  resultTopicFromNodes, resultComplementFromNodes, makeChangeIntoPredicate,
+  wrapChangeIntoPredicateSubspans,
+} = resultComplementModule.createResultComplementDetectors({
+  categorySubspanFor, construction, flattenSurface, isToken, nodeCanFillSlot,
+  nominalComplementFromNodes, resultFramePartClone, templateDerivedSlots, traceInfo,
+  withoutIgnorableSpaceText,
+});
 
 
 
 
 
-function makeChangeIntoPredicate(changeNode, complement) {
-  const children = [
-    resultFramePartClone(changeNode, {
-      label: "doing",
-      pos: "verb",
-      syntax: "change_into_verb result_change_verb",
-      slots: ["change_verb", "action_verb", "main_verb", "predicate"],
-      reason: "變成 is the change-into predicate head inside a bounded change-result frame.",
-    }),
-    complement,
-  ];
-  return construction("ChangeIntoPredicate", "變成", children, {
-    note: "Bounded change-into predicate: 變成 + result complement.",
-    slots: templateDerivedSlots("ChangeIntoPredicate", children),
-    trace: traceInfo("generative_template", {
-      construction_type: "ChangeIntoPredicate",
-      template: ["change_verb!", "result_complement!"],
-      assigned_slots: ["change_verb", "result_complement"],
-      surfaces: children.map((node) => flattenSurface(node)),
-      subspan: true,
-    }),
-  });
-}
 
-function wrapChangeIntoPredicateSubspans(nodes) {
-  const result = [];
-  let i = 0;
-  while (i < nodes.length) {
-    if (isToken(nodes[i], "變成") && isToken(nodes[i + 1], "點")) {
-      result.push(makeChangeIntoPredicate(nodes[i], nodes[i + 1]));
-      i += 2;
-      continue;
-    }
-    result.push(nodes[i]);
-    i += 1;
-  }
-  return result;
-}
+
+
+
 
 
 function bridgeFramePartClone(node, overrides = {}) {
@@ -3525,201 +3393,19 @@ function compositionPartClone(node, overrides = {}) {
   });
 }
 
-function durativeAspectCompositionFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  const compact = withoutIgnorableSpaceText(bareCore);
-  if (compact.length !== 3 || !nodeCanFillSlot(compact[0], "action_verb") || flattenSurface(compact[1]) !== "住") return null;
-  if (!nodeCanFillSlot(compact[2], "object") && !nodeCanFillSlot(compact[2], "head_noun") && !nodeCanFillSlot(compact[2], "np")) return null;
-  const marker = parserInactiveTokenClone(firstToken(compact[1]) || token("住"), {
-    label: "func", pos: "function", syntax: "durative_aspect", slots: ["durative_aspect", "aspect_marker"], jyutping: "zyu6",
-    note: "住 marks the continuing wearing/resultant state, distinct from progressive 緊.",
-    reason: "住 is interpreted as durative aspect only inside the licensed action + 住 + object pattern.",
-  });
-  const children = [compact[0], marker, compact[2], ...particles];
-  return construction("DurativeVP", "DurativeVP", children, {
-    slots: constructionSlotsByType("DurativeVP", children),
-    trace: traceInfo("generative_template", {
-      construction_type: "DurativeVP", template_family: "generative_template",
-      template: ["action_verb!", "durative_aspect!", "object!", "particle?"],
-      assigned_slots: ["action_verb", "durative_aspect", "object", ...particles.map(() => "particle")],
-      surfaces: children.map(flattenSurface), aspect_type: "durative_continuing_state",
-      not_claims: ["not_progressive_event", "not_global_住_lexicalization"],
-    }),
-  });
-}
+const createAspectCompositionDetectors = require("./parser/detectors/aspect/composition");
+const {
+  durativeAspectCompositionFallback, perfectiveResultCompositionFallback,
+  restorativeRepetitiveComplementFallback, incompatibleAspectCompositionMalformedCandidate,
+} = createAspectCompositionDetectors({
+  classifierObjectNPFromNodes, cleanSlots, compositionPartClone, construction,
+  constructionSlotsByType, firstToken, flattenSurface, isToken, nodeCanFillSlot,
+  nodeSlots, parserInactiveTokenClone, templateDerivedSlots, traceInfo, withoutIgnorableSpaceText, withoutTrailingParticles,
+});
 
-function perfectiveResultCompositionFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  const compact = withoutIgnorableSpaceText(bareCore);
-  let cursor = 0;
-  const subject = compact[cursor] && nodeCanFillSlot(compact[cursor], "subject") ? compact[cursor++] : null;
-  if (compact.length - cursor !== 4) return null;
-  const [action, result, aspect, object] = compact.slice(cursor);
-  if (!nodeCanFillSlot(action, "action_verb") || !nodeCanFillSlot(aspect, "perfective_aspect")) return null;
-  if (!nodeCanFillSlot(object, "object") && !nodeCanFillSlot(object, "head_noun") && !nodeCanFillSlot(object, "np")) return null;
-  const resultSurface = flattenSurface(result);
-  if (!nodeCanFillSlot(result, "completion_marker") && resultSurface !== "好") return null;
-  const innerType = resultSurface === "好" ? "ResultComplementVP" : "CompletionVP";
-  const inner = construction(innerType, resultSurface === "好" ? "ResultVP" : "CompletionVP", [
-    action,
-    compositionPartClone(result, {
-      label: resultSurface === "好" ? "how" : "func",
-      syntax: resultSurface === "好" ? "result_state_complement" : "completion_result_complement",
-      slots: resultSurface === "好" ? ["result_complement"] : ["completion_marker", "result_complement"],
-      reason: `${resultSurface} is the overt result/phase complement formed with the action before perfective 咗 scopes over the complex predicate.`,
-    }),
-  ], {
-    slots: cleanSlots([innerType === "CompletionVP" ? "completion_vp" : "result_complement_vp", "result_complement", "vp", "action_vp", "predicate"]),
-    note: "Inner result/phase predicate formed before perfective aspect.",
-    trace: traceInfo("generative_template", {
-      construction_type: innerType,
-      template_family: "generative_template",
-      template: ["action_verb!", "result_or_phase_complement!"],
-      assigned_slots: ["action_verb", "result_complement"],
-      surfaces: [flattenSurface(action), resultSurface],
-      aspect_scope_status: "inside_perfective_scope",
-    }),
-  });
-  const perfective = construction("PerfectiveVP", "PerfectiveVP", [
-    inner,
-    compositionPartClone(aspect, { label: "func", syntax: "perfective_aspect", slots: ["perfective_aspect", "aspect_marker"] }),
-    object,
-    ...particles,
-  ], {
-    slots: cleanSlots(["perfective_vp", "vp", "action_vp", "predicate", "perfective_aspect", "object"]),
-    note: "Perfective aspect scopes over an already formed result/phase predicate while surface order remains unchanged.",
-    trace: traceInfo("generative_template", {
-      construction_type: "PerfectiveVP",
-      template_family: "generative_template",
-      template: ["result_or_phase_vp!", "perfective_aspect!", "object!", "particle?"],
-      assigned_slots: ["predicate", "perfective_aspect", "object", ...particles.map(() => "particle")],
-      surfaces: [flattenSurface(inner), flattenSurface(aspect), flattenSurface(object), ...particles.map(flattenSurface)],
-      aspect_scope_status: "perfective_over_result_complex",
-      not_claims: ["not_flat_aspect_stack", "not_hidden_object"],
-    }),
-  });
-  if (!subject) return perfective;
-  const children = [subject, perfective];
-  return construction("SubjectPredicateClause", "SubjPred", children, {
-    slots: templateDerivedSlots("SubjectPredicateClause", children),
-    trace: traceInfo("generative_template", { construction_type: "SubjectPredicateClause", template_family: "generative_template", template: ["subject!", "predicate!"], assigned_slots: ["subject", "predicate"], surfaces: children.map(flattenSurface) }),
-  });
-}
 
-function restorativeRepetitiveComplementFallback(core) {
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  const compact = withoutIgnorableSpaceText(bareCore);
-  if (compact.length < 3 || !nodeCanFillSlot(compact[0], "action_verb") || !isToken(compact[1], "返")) return null;
-  const action = compact[0];
-  const marker = (reading) => compositionPartClone(compact[1], {
-    label: "func",
-    syntax: reading === "restorative"
-      ? "restorative_complement_marker"
-      : "repetitive_resumptive_complement_marker",
-    slots: reading === "restorative"
-      ? ["verb_complement", "restorative_complement"]
-      : ["verb_complement", "repetitive_complement"],
-    reason: reading === "restorative"
-      ? "返 marks restoration of a prior or expected state here; it is not literal return motion."
-      : "返 marks repetition or resumption of the overt action here; it is not literal return motion.",
-  });
-  const thirdSurface = flattenSurface(compact[2]);
-  if (thirdSurface === "好" && compact.length === 3) {
-    const children = [action, marker("restorative"), compositionPartClone(compact[2], { label: "how", syntax: "restored_result_state", slots: ["result_complement"] }), ...particles];
-    return construction("RestorativeComplementVP", "RestoreVP", children, { slots: constructionSlotsByType("RestorativeComplementVP", children), trace: traceInfo("generative_template", { construction_type: "RestorativeComplementVP", template_family: "generative_template", template: ["action_verb!", "restorative_complement!", "result_state!", "particle?"], assigned_slots: ["action_verb", "restorative_complement", "result_complement", ...particles.map(() => "particle")], surfaces: children.map(flattenSurface), return_reading: "restorative", not_claims: ["not_literal_return_motion", "not_generic_aspect"] }) });
-  }
-  if (compact.length === 3 && (nodeCanFillSlot(compact[2], "frequency_quantity") || /(^|\s)frequency_quantity(\s|$)/.test((firstToken(compact[2]) || {}).syntax || ""))) {
-    const children = [action, marker("repetitive"), compact[2], ...particles];
-    return construction("RepetitiveComplementVP", "RepeatVP", children, { slots: constructionSlotsByType("RepetitiveComplementVP", children), trace: traceInfo("generative_template", { construction_type: "RepetitiveComplementVP", template_family: "generative_template", template: ["action_verb!", "repetitive_complement!", "frequency_quantity!", "particle?"], assigned_slots: ["action_verb", "repetitive_complement", "frequency_quantity", ...particles.map(() => "particle")], surfaces: children.map(flattenSurface), return_reading: "repetitive_or_resumptive", not_claims: ["not_literal_return_motion", "not_generic_aspect"] }) });
-  }
-  if (compact.length === 4) {
-    const object = classifierObjectNPFromNodes(compact.slice(2));
-    if (object) {
-      const actionSlots = new Set(nodeSlots(action));
-      const actionSyntax = String((firstToken(action) || {}).syntax || "");
-      const licensesResumptiveObjectReading = actionSlots.has("consumption_verb")
-        || actionSyntax.includes("chain_select_perception")
-        || actionSyntax.includes("chain_select_discourse_content");
-      const reading = licensesResumptiveObjectReading ? "repetitive" : "underdetermined";
-      const returnPart = reading === "repetitive"
-        ? marker("repetitive")
-        : compositionPartClone(compact[1], {
-          label: "func",
-          syntax: "return_or_resumptive_complement_marker",
-          slots: ["verb_complement"],
-          reason: "返 is a non-finite complement marker here. Available evidence does not force literal motion or a repetitive reading.",
-        });
-      const children = [action, returnPart, object, ...particles];
-      const type = reading === "repetitive" ? "RepetitiveComplementVP" : "VerbComplementVP";
-      const label = reading === "repetitive" ? "RepeatVP" : "VerbCompVP";
-      return construction(type, label, children, {
-        slots: constructionSlotsByType(type, children),
-        trace: traceInfo("generative_template", {
-          construction_type: type,
-          template_family: "generative_template",
-          template: ["action_verb!", reading === "repetitive" ? "repetitive_complement!" : "return_or_resumptive_complement!", "object!", "particle?"],
-          assigned_slots: ["action_verb", reading === "repetitive" ? "repetitive_complement" : "verb_complement", "object", ...particles.map(() => "particle")],
-          surfaces: children.map(flattenSurface),
-          return_reading: reading === "repetitive" ? "repetitive_or_resumptive" : "context_underdetermined_return_or_resumptive",
-          not_claims: reading === "repetitive"
-            ? ["not_literal_return_motion", "not_generic_aspect"]
-            : ["not_forced_literal_return_motion", "not_forced_repetitive_reading", "not_generic_aspect"],
-        }),
-      });
-    }
-  }
-  return null;
-}
 
-function incompatibleAspectCompositionMalformedCandidate(core) {
-  if (!core || core.length < 3) return null;
-  const { core: bareCore, particles } = withoutTrailingParticles(core);
-  let cursor = 0;
-  const subject = bareCore[cursor] && nodeCanFillSlot(bareCore[cursor], "subject") ? bareCore[cursor++] : null;
-  const action = bareCore[cursor++];
-  if (!action || !nodeCanFillSlot(action, "action_verb")) return null;
-  const firstMarker = bareCore[cursor++];
-  const secondMarker = bareCore[cursor++];
-  if (!firstMarker || !secondMarker) return null;
 
-  let malformedSubtype = "";
-  let problem = "";
-  let expected = [];
-  if (nodeCanFillSlot(firstMarker, "perfective_aspect") && nodeCanFillSlot(secondMarker, "progressive_aspect")) {
-    malformedSubtype = "incompatible_perfective_progressive_stack";
-    problem = "Perfective 咗 and progressive 緊 are stacked in an incompatible order on one predicate.";
-    expected = ["one_licensed_aspect_layer", "separate_clause_or_repair"];
-  } else if ((isToken(firstMarker, "得") || isToken(firstMarker, "唔")) && nodeCanFillSlot(secondMarker, "perfective_aspect")) {
-    malformedSubtype = isToken(firstMarker, "得") ? "potential_marker_followed_by_perfective" : "potential_negator_followed_by_perfective";
-    problem = "Potential 得/唔 is followed by perfective 咗 instead of an overt result complement.";
-    expected = ["overt_result_complement_after_potential_marker", "ordinary_perfective_vp_without_potential_marker"];
-  } else if (["嚟", "去"].includes(flattenSurface(firstMarker)) && ["入", "出", "返", "上", "落"].includes(flattenSurface(secondMarker))) {
-    malformedSubtype = "deictic_marker_not_outermost";
-    problem = "Final deictic 嚟/去 precedes a path or return component instead of occupying the outer edge of the directional complex.";
-    expected = ["path_or_return_before_final_deictic_marker", "separate_clause_or_repair"];
-  } else {
-    return null;
-  }
-
-  const children = [...(subject ? [subject] : []), action, firstMarker, secondMarker, ...bareCore.slice(cursor), ...particles];
-  return construction("MalformedCandidate", "Malformed", children, {
-    slots: cleanSlots(["malformed_candidate", "needs_review", "predicate", "problem_span", "action_verb", subject ? "subject" : ""]),
-    note: "Review-bearing incompatible aspect/potential composition; all visible material is preserved without repairing the learner input.",
-    trace: traceInfo("special_ambiguity_rule", {
-      construction_type: "MalformedCandidate",
-      malformed_family: "aspect_result_potential_composition",
-      malformed_subtype: malformedSubtype,
-      template: ["subject?", "action_verb!", "incompatible_marker_1!", "incompatible_marker_2!", "remainder?", "particle?"],
-      assigned_slots: [...(subject ? ["subject"] : []), "action_verb", "incompatible_marker_1", "incompatible_marker_2", ...(bareCore.slice(cursor).length ? ["remainder"] : []), ...particles.map(() => "particle")],
-      surfaces: children.map((node) => flattenSurface(node)),
-      problem,
-      expected_repairs: expected,
-      semantic_review_flags: ["malformed_candidate_parse", "incompatible_aspect_or_potential_order"],
-      not_claims: ["not_clean_aspect_stack", "not_hidden_result_complement", "not_silent_input_repair"],
-      reason: "Aspect, potential, and result layers must compose in a licensed order; incompatible visible markers remain review-bearing."
-    })
-  });
-}
 
 function environmentalPredicateParts(core = []) {
   const compact = withoutIgnorableSpaceText(core || []);
@@ -4382,87 +4068,16 @@ function nominalMeasurePredicateFallback(core) {
 
 
 
-function postverbalZoPerfectiveFromRawNodes(nodes = []) {
-  const compact = withoutIgnorableSpaceText(nodes || []);
-  if (compact.length < 3) return null;
-  const subjectOffset = nodeCanFillSlot(compact[0], "subject") ? 1 : 0;
-  if (compact.length - subjectOffset < 3) return null;
-  const action = compact[subjectOffset];
-  const aspect = compact[subjectOffset + 1];
-  if (!nodeCanFillSlot(action, "action_verb") || !nodeCanFillSlot(aspect, "perfective_aspect")) return null;
-  const objectNodes = compact.slice(subjectOffset + 2);
-  if (objectNodes.length < 2) return null;
-  const objectNode = compositionalNpSubspanFor(objectNodes)
-    || (objectNodes.length === 1 && nodeCanLicenseEvidenceGatedObject(objectNodes[0]) ? objectNodes[0] : null);
-  if (!objectNode || !nodeCanLicenseEvidenceGatedObject(objectNode)) return null;
-  const children = [action, aspect, objectNode];
-  const perfective = construction("PostverbalZoPerfectiveVP", "PerfectiveVP", children, {
-    slots: cleanSlots(["perfective_vp", "vp", "action_vp", "predicate", "action_verb", "perfective_aspect", "object", "np"]),
-    note: "Compositional postverbal 咗 perfective with an overt licensed NP object. NP assembly is independent parser infrastructure and does not broaden the construction's linguistic status.",
-    trace: traceInfo("generative_template", {
-      construction_type: "PostverbalZoPerfectiveVP",
-      template_family: "generative_template",
-      template: ["action_verb!", "perfective_aspect!", "licensed_np_object!"],
-      assigned_slots: ["action_verb", "perfective_aspect", "object"],
-      surfaces: children.map((node) => flattenSurface(node)),
-      object_np_license_status: nodeNpLicenseStatus(objectNode),
-      object_np_construction: objectNode.kind === "construction" ? objectNode.type : "bare_nominal_token",
-      hidden_object_inserted: false,
-      evidence_scope_unchanged: true,
-      reason: "v0.5.184 composes the complete postverbal object as a reusable NP before broad VP subspan wrapping can consume only its first token.",
-    }),
-  });
-  if (!subjectOffset) return [perfective];
-  const clause = templateConstructionFor([compact[0], perfective], ["SubjectPredicateClause"]);
-  return clause ? [clause] : [compact[0], perfective];
-}
+const createPostverbalZoDetectors = require("./parser/detectors/aspect/postverbal-zo");
+const {
+  postverbalZoPerfectiveFromRawNodes, postverbalZoPerfectiveFromWrappedNodes,
+} = createPostverbalZoDetectors({
+  cleanSlots, compositionalNpSubspanFor, construction, flattenSurface,
+  nodeCanFillSlot, nodeCanLicenseEvidenceGatedObject, nodeNpLicenseStatus,
+  templateConstructionFor, traceInfo, withoutIgnorableSpaceText,
+});
 
-function postverbalZoPerfectiveFromWrappedNodes(nodes = []) {
-  const compact = withoutIgnorableSpaceText(nodes || []);
-  if (compact.length < 2) return null;
-  for (let index = 0; index < compact.length - 1; index += 1) {
-    const first = compact[index];
-    const objectNode = compact[index + 1];
-    if (!nodeCanLicenseEvidenceGatedObject(objectNode)) continue;
-    let action = null;
-    let aspect = null;
-    if (first && first.kind === "construction" && first.type === "PerfectiveVP") {
-      const children = withoutIgnorableSpaceText(first.children || []);
-      if (children.length === 2 && nodeCanFillSlot(children[0], "action_verb") && nodeCanFillSlot(children[1], "perfective_aspect")) {
-        [action, aspect] = children;
-      }
-    } else if (index + 2 < compact.length
-        && nodeCanFillSlot(first, "action_verb")
-        && nodeCanFillSlot(compact[index + 1], "perfective_aspect")
-        && nodeCanLicenseEvidenceGatedObject(compact[index + 2])) {
-      action = first;
-      aspect = compact[index + 1];
-    }
-    if (!action || !aspect) continue;
-    const actualObject = first && first.kind === "construction" ? objectNode : compact[index + 2];
-    const consumed = first && first.kind === "construction" ? 2 : 3;
-    if (index + consumed !== compact.length) continue;
-    const children = [action, aspect, actualObject];
-    const perfective = construction("PostverbalZoPerfectiveVP", "PerfectiveVP", children, {
-      slots: cleanSlots(["perfective_vp", "vp", "action_vp", "predicate", "action_verb", "perfective_aspect", "object", "np"]),
-      note: "Compositional postverbal 咗 perfective with an overt licensed NP object. NP assembly is independent parser infrastructure and does not broaden the construction's linguistic status.",
-      trace: traceInfo("generative_template", {
-        construction_type: "PostverbalZoPerfectiveVP",
-        template_family: "generative_template",
-        template: ["action_verb!", "perfective_aspect!", "licensed_np_object!"],
-        assigned_slots: ["action_verb", "perfective_aspect", "object"],
-        surfaces: children.map((node) => flattenSurface(node)),
-        object_np_license_status: nodeNpLicenseStatus(actualObject),
-        object_np_construction: actualObject.kind === "construction" ? actualObject.type : "bare_nominal_token",
-        hidden_object_inserted: false,
-        evidence_scope_unchanged: true,
-        reason: "v0.5.184 recomposes V+咗 with a reusable licensed NP after NP-internal parsing, instead of enumerating complete object strings.",
-      }),
-    });
-    return [...compact.slice(0, index), perfective, ...compact.slice(index + consumed)];
-  }
-  return null;
-}
+
 
 const createDefinitionCopularDetectors = require("./parser/detectors/definition/copular-relations");
 const {
